+++ app/controllers/application_controller.rb
@@ -6,6 +6,10 @@ class ApplicationController < ActionController::Base
 
   helper_method :current_user 
 
+  def snake_case_params
+    request.parameters.deep_transform_keys!(&:underscore)
+  end
+
   def confirm_password
     if (params["password"] != params["passwordConfirmation"])
       render(json: { error: "Password and Password Confirmation do not match" }, status: 401)
@@ -13,23 +17,6 @@ class ApplicationController < ActionController::Base
     else
       return true
     end
-  end
-
-  def check_patient_code
-    if (check_patient_activation_code)
-      render(json: { validCode: true }, status: 200)
-    else
-      if (@temp_account)
-        render(json: { validCode: false }, status: 401)
-      else
-        render(json: { validCode: false }, status: 422)
-      end
-    end
-  end
-
-  def check_patient_activation_code
-    @temp_account = TempAccount.where(phone_number: params["phoneNumber"]).first()
-    return (@temp_account && BCrypt::Password.new(@temp_account.code_digest) == params["activationCode"])
   end
 
   def get_locales
