+++ app/controllers/patient_controller.rb
@@ -35,6 +35,7 @@ class PatientController < UserController
     else
       @current_user.contact_tracing = ContactTracing.create!(number_of_contacts: params[:numberContacts], contacts_tested: params[:contactsTested], patient_id: @current_user.id)
     end
+
     @current_user.add_photo_day
     @current_user.update(gender: params[:gender], age: params[:age], app_start: DateTime.now(), status: "Active", gender_other: params[:genderOther] || nil)
     @current_user.save!
