+++ app/policies/application_policy.rb
@@ -46,4 +46,5 @@ class ApplicationPolicy
       scope.all
     end
   end
+
 end
